requirements.txt
opencv-python-headless
tensorflow
wheel
nbresult
colorama
ipdb
ipykernel
yapf
matplotlib
pygeohash
pytest
seaborn
numpy
pandas
scipy
scikit-learn
google-cloud-bigquery
google-cloud-storage
db-dtypes
pyarrow
streamlit
mlflow==2.2.1

python-dotenv
prefect
psycopg2-binary

# API
fastapi
pytz
uvicorn
# tests
httpx
pytest-asyncio

